requirements.txt
crewai
crewai-tools[singlestore]
python-dotenv